flask/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.7
redis==4.5.5
psycopg2-binary==2.9.5
pymongo==4.3.3
